part_4-web-apps/part-i-python-apis/10_api-authentication-and-authorization/03_jwt-generator/pyproject.toml
[tool.poetry]
name = "jwt-generator"
version = "0.1.0"
description = ""
authors = ["Sergio F. Gonzalez <sergio.f.gonzalez@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyjwt = "^2.8.0"
cryptography = "^42.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
